Bump Ruby and Gems
When I pulled the site down just now and followed the instructions to
get it running, nothing worked. I tried Ruby 2.7 and then 2.6 and just
got Forbidden and Not Found errors from Webrick.

Given that the site was running super old versions of everything, I
figured it might be wise to bundle update to more recent versions. I
updated to Jekyll 3.8.x because 4.0.0 had some deprecations I didn't
want to figure out around Kramdown. The configuration options I removed
weren't needed anymore and the Rakefile was updated to reflect the new
way of invoking jekyll to serve the site (which fixed some deprecation
warnings.)

Ruby 2.7 spat out tons of warning because of the keyword argument stuff,
so I just used the latest 2.6 release.

With this commit, the instructions in the README now work for me and
we're not running ancient versions of things.